QMAPS-1649 pr fixes
